+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 1. [#252](https://github.com/influxdata/influxdb-client-java/pull/252): Spring auto-configuration works even without `influxdb-client-reactive` [spring]
 1. [#254](https://github.com/influxdata/influxdb-client-java/pull/254): Avoid reading entire query response into bytes array
 
+### Deprecates
+1. [#255](https://github.com/influxdata/influxdb-client-java/pull/255): `InfluxDBClient#getWriteApi()` instead use `InfluxDBClient#makeWriteApi()`
+
 ## 3.1.0 [2021-07-27]
 
 ### Breaking Changes